fix casting of int

+++ djangoapi/realiumapi/api/views.py
@@ -256,7 +256,7 @@ class EventView(generics.GenericAPIView):
 
         if request.data['eventType']=='SALE':
             numTokens = int(request.data['quantity'])
-            tokensToBeSold = self.token_model.objects.filter(property__propertyId=int(request.data['property']),owner__realiumUserId=int(request.data['tokenOwner'], listedPrice=int(request.data['listedPrice'])),listed=True)[:numTokens]
+            tokensToBeSold = self.token_model.objects.filter(property__propertyId=int(request.data['property']),owner__realiumUserId=int(request.data['tokenOwner'], listedPrice=float(request.data['listedPrice'])),listed=True)[:numTokens]
             txNFTId = str('')
             txAvaxId = str('')
             if len(tokensToBeSold)>0:
